Mock crate separation from binaries (#375)
* mock_macro: Add initial.

* mock_macro: Add initialize_database.

* actor: Pilot mock_macro usage.

* mock_macro: Add needed imports for integration tests.

* tests: Use new mock_macro in integration tests.

* mock_macro: Add create_actors.

* mock_macro: Use create_actors everywhere and delete create_actors_grpc.

* mock_macro: Use initialize_database everywhere.

* mock_macro: Use create_test_config_with_thread_name in everywhere.

* test_utils: Rename mock_macro.

* cargo: Delete bin entry of `all_servers`.

* test_utils: Remove create_database and drop_database to add them in initialize_database.

* test_utils: Use get_postgresql_url in initialize_database.

* JsonRPC endpoint removals (#377)

* jsonrpc: Initial remove.

* operator: Remove old rpc test.

* verifier: Convert functions to pub.

* tests: Delete old file.

* cargo: Remove unused features for jsonrpsee.
