+++ Cargo.toml
@@ -10,7 +10,7 @@ serde_json = "1.0.128"
 thiserror = "1.0.64"
 tracing = { version = "0.1.40", default-features = false }
 tracing-subscriber = { version = "0.3.18", features = ["json"] }
-jsonrpsee = "0.22.5"
+jsonrpsee = { version = "0.22.5", default-features = false }
 async-trait = "0.1.83"
 clap = "4.5.20"
 toml = "0.8.19"
