+++ core/Cargo.toml
@@ -21,7 +21,7 @@ thiserror = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 tokio = { workspace = true, features = ["full"] }
-jsonrpsee = { workspace = true, features = ["server", "http-client", "macros"] }
+jsonrpsee = { workspace = true, features = ["http-client"] }
 async-trait = { workspace = true }
 futures = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
@@ -49,11 +49,6 @@ testing = []
 [[bin]]
 name = "server"
 path = "src/bin/server.rs"
-
-[[bin]]
-name = "all_servers"
-path = "src/bin/all_servers.rs"
-required-features = ["testing"]
 
 [[bin]]
 name = "config_generator"
