+++ core/src/header_chain_prover/blockgazer.rs
@@ -203,15 +203,19 @@ impl HeaderChainProver {
 #[cfg(test)]
 mod tests {
     use crate::{
+        config::BridgeConfig, database::Database, initialize_database, utils::initialize_logger,
+    };
+    use crate::{
+        create_test_config_with_thread_name,
         extended_rpc::ExtendedRpc,
         header_chain_prover::{
             blockgazer::{BlockFetchStatus, BATCH_DEEPNESS, BATCH_DEEPNESS_SAFETY_BARRIER},
             HeaderChainProver,
         },
-        mock::database::create_test_config_with_thread_name,
     };
     use bitcoin::BlockHash;
     use bitcoincore_rpc::RpcApi;
+    use std::{env, thread};
 
     async fn mine_and_save_blocks(prover: &HeaderChainProver, height: u64) -> Vec<BlockHash> {
         let mut fork_block_hashes = Vec::new();
@@ -253,7 +257,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn check_for_new_blocks_uptodate() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -287,7 +291,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn check_for_new_blocks_fallen_behind_single() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -328,7 +332,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn check_for_new_blocks_fallen_behind_multiple() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -359,7 +363,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn check_for_new_blocks_fork_and_mine_new() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -391,7 +395,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn sync_blockchain_single_block() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -425,7 +429,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn sync_blockchain_multiple_blocks() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
@@ -459,7 +463,7 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn sync_blockchain_multiple_blocks_with_fork() {
-        let config = create_test_config_with_thread_name("test_config.toml", None).await;
+        let config = create_test_config_with_thread_name!("test_config.toml", None);
         let rpc = ExtendedRpc::new(
             config.bitcoin_rpc_url.clone(),
             config.bitcoin_rpc_user.clone(),
