--- core/src/mock/mod.rs
@@ -1,8 +1,0 @@
-//! # Mock Interfaces
-//!
-//! This module includes mock interfaces for tests. There are also some common
-//! elements for unit and integration tests.
-
-pub mod common;
-pub mod database;
-pub mod env;
