+++ core/src/watchtower.rs
@@ -72,17 +72,27 @@ impl Watchtower {
 
 #[cfg(test)]
 mod tests {
+    use crate::utils::initialize_logger;
+    use crate::watchtower::Watchtower;
     use crate::{
-        mock::database::create_test_config_with_thread_name, servers::create_actors_grpc,
-        watchtower::Watchtower,
+        config::BridgeConfig,
+        database::Database,
+        errors::BridgeError,
+        extended_rpc::ExtendedRpc,
+        initialize_database,
+        servers::{
+            create_aggregator_grpc_server, create_operator_grpc_server,
+            create_verifier_grpc_server, create_watchtower_grpc_server,
+        },
     };
+    use crate::{create_actors, create_test_config_with_thread_name};
+    use std::{env, thread};
 
     #[tokio::test]
     #[serial_test::serial]
     async fn new_watchtower() {
-        let mut config = create_test_config_with_thread_name("test_config.toml", None).await;
-        let (verifiers, operators, _, _should_not_panic) =
-            create_actors_grpc(config.clone(), 1).await;
+        let mut config = create_test_config_with_thread_name!("test_config.toml", None);
+        let (verifiers, operators, _, _should_not_panic) = create_actors!(config.clone(), 1);
 
         config.verifier_endpoints = Some(
             verifiers
@@ -103,8 +113,8 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn get_winternitz_public_keys() {
-        let mut config = create_test_config_with_thread_name("test_config.toml", None).await;
-        let (verifiers, operators, _, _watchtowers) = create_actors_grpc(config.clone(), 2).await;
+        let mut config = create_test_config_with_thread_name!("test_config.toml", None);
+        let (verifiers, operators, _, _watchtowers) = create_actors!(config.clone(), 2);
 
         config.verifier_endpoints = Some(
             verifiers
