+++ core/tests/taproot.rs
@@ -7,13 +7,16 @@ use clementine_core::actor::Actor;
 use clementine_core::builder::transaction::TxHandler;
 use clementine_core::builder::{self};
 use clementine_core::extended_rpc::ExtendedRpc;
-use clementine_core::mock::database::create_test_config_with_thread_name;
 use clementine_core::utils::{handle_taproot_witness_new, SECP};
+use clementine_core::{config::BridgeConfig, database::Database, utils::initialize_logger};
+use std::{env, thread};
+
+mod common;
 
 #[tokio::test]
 #[serial_test::serial]
 async fn create_address_and_transaction_then_sign_transaction() {
-    let config = create_test_config_with_thread_name("test_config.toml", None).await;
+    let config = create_test_config_with_thread_name!("test_config.toml", None);
     let rpc = ExtendedRpc::new(
         config.bitcoin_rpc_url,
         config.bitcoin_rpc_user,
